requirements.txt
pyspark>=3.4.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
pytest>=7.4.0
pandas>=2.0.0
